Added a WindowHandler class
